requirements.txt
einops
fire
jax
jaxlib
opax
pax3
Pillow
tensorflow_datasets